requirements.txt
pytest
autopep8
flake8
redis
torch
transformers
pillow
faiss-cpu